requirements.txt
pyTelegramBotAPI==4.29.1
google-genai
functions-framework>=3.0.0